fix: 소수점 관련 제대로 토글되지 않는 문제 수정

+++ Calculator/Calculator/Controller/CalculatorViewController.swift
@@ -50,11 +50,10 @@ final class CalculatorViewController: UIViewController {
         }
     }
 
-    // TODO: - Int 말고 Double
     @IBAction func signToggleDidTap(_ sender: UIButton) {
         guard nil != sender.currentTitle,
               let currentNumberString = entryNumberLabel.text,
-              let currentNumber = Int(currentNumberString) else { return }
+              let currentNumber = Double(currentNumberString) else { return }
         entryNumberLabel.text = String(currentNumber * -1)
     }
